Adding resource requests and limits to deployment
According to https://help.sonatype.com/repomanager3/system-requirements#SystemRequirements-InstanceSizingProfiles 4GB is the recommended minimum RAM for a small installation of Nexus Repomanager 3. It is also for that size of RAM the image is tuned by default https://hub.docker.com/r/sonatype/nexus3#notes

When setting the request to the same as the limit I'm taking into account that lucene utilises not only userland memory, but also OS buffers.